% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaData}
\alias{brcaData}
\title{FUNCTION_TITLE}
\format{
An object of class \code{data.frame} with 54675 rows and 47 columns.
}
\usage{
brcaData
}
\description{
FUNCTION_DESCRIPTION
}
\keyword{datasets}
